requirements.txt
# SubVivienda IA — Dependencias
# Compatible con Windows + Python 3.10 en entorno virtual limpio

# API Backend
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.9.0
python-multipart==0.0.9

# Frontend
streamlit==1.35.0

# Data
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
requests==2.31.0

# Machine Learning
scikit-learn==1.4.2
xgboost==2.0.3

# Deep Learning
# tensorflow-cpu 2.15.0 resuelve el conflicto de typing-extensions
# con FastAPI, Streamlit y Pydantic en Windows Python 3.10
tensorflow-cpu==2.15.0

# Explicabilidad y NLP
shap==0.45.1
gensim==4.3.2

# Visualizacion
plotly==5.22.0
